Telegram Bot/requirements.txt
pyTelegramBotAPI
python-whois
python-dotenv
requests 